+++ src/com/javaex/ex04/Rectangle.java
@@ -10,7 +10,7 @@ public class Rectangle extends Shape implements Resizeable {
 	public Rectangle() {
 		super();
 	}
-	public Rectangle(int width, int height) {
+	public Rectangle(double width, double height) {
 		super();
 		this.width = width;
 		this.height = height;
@@ -20,13 +20,13 @@ public class Rectangle extends Shape implements Resizeable {
 	public double getWidth() {
 		return width;
 	}
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 	public double getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
